+++ AddOns/Decursive/Dcr_DIAG.lua
@@ -1,7 +1,7 @@
 --[[
     This file is part of Decursive.
 
-    Decursive (v 2.8.3-19-gef0d480) add-on for World of Warcraft UI
+    Decursive (v 2.8.3-25-g9cacdb5) add-on for World of Warcraft UI
     Copyright (C) 2006-2026 John Wellesz (Decursive AT 2072productions.com) ( http://www.2072productions.com/to/decursive.php )
 
     Decursive is free software: you can redistribute it and/or modify
@@ -24,7 +24,7 @@
     Decursive is distributed in the hope that it will be useful,
     but WITHOUT ANY WARRANTY.
 
-    This file was last updated on 2026-08-28T12:24:50Z
+    This file was last updated on 2026-09-02T10:28:21Z
 --]]
 -------------------------------------------------------------------------------
 
@@ -118,7 +118,7 @@ local DebugTextTable    = T._DebugTextTable;
 local Reported          = {};
 
 local UNPACKAGED = "@pro" .. "ject-version@";
-local VERSION = "2.8.3-19-gef0d480";
+local VERSION = "2.8.3-25-g9cacdb5";
 
 if not T._LoadedFiles then
     T._LoadedFiles = {};
@@ -174,6 +174,8 @@ T._LoadOrderedFiles = { -- {{{
 
     "Dcr_DebuffsFrame.lua",
     "Dcr_DebuffsFrame.xml",
+
+    "Dcr_12_1_Sounds.lua",
 
     "Dcr_LiveList.lua",
     "Dcr_LiveList.xml",
@@ -340,7 +342,7 @@ do
         local dbclud = T.Dcr.Status and T.Dcr.Status.delayedUnDebuffOccurences or -1
 
 
-        DebugHeader = ("%s\n2.8.3-19-gef0d480  %s(%s)  CT: %0.4f D: %s %s %s DTl: %d DE: %d nDrE: %d Embeded: %s W: %d (LA: %d TAMU: %d) TA: %d NDRTA: %d BUIE: %d dbc: [d:%d-%d, u:%d-%d] TI: [dc:%d, lc:%d, y:%d, LEBY:%d, LB:%d, TTE:%u] (%s, %s, %s, %s)"):format(instructionsHeader, -- "%s\n
+        DebugHeader = ("%s\n2.8.3-25-g9cacdb5  %s(%s)  CT: %0.4f D: %s %s %s DTl: %d DE: %d nDrE: %d Embeded: %s W: %d (LA: %d TAMU: %d) TA: %d NDRTA: %d BUIE: %d dbc: [d:%d-%d, u:%d-%d] TI: [dc:%d, lc:%d, y:%d, LEBY:%d, LB:%d, TTE:%u] (%s, %s, %s, %s)"):format(instructionsHeader, -- "%s\n
         tostring(DC.MyClass), tostring(UnitLevel("player") or "??"), NiceTime(), date(), GetLocale(), -- %s(%s)  CT: %0.4f D: %s %s
         BugGrabber and "BG" .. (T.BugGrabber and "e" or "") or "NBG", -- %s
         #DebugTextTable / 2, -- DTl: %d
@@ -1216,4 +1218,4 @@ do
     end
 end
 
-T._LoadedFiles["Dcr_DIAG.lua"] = "2.8.3-19-gef0d480";
+T._LoadedFiles["Dcr_DIAG.lua"] = "2.8.3-25-g9cacdb5";